+++ cmake/modules/DuneDoxygen.cmake
@@ -109,7 +109,7 @@ set_property(GLOBAL PROPERTY DUNE_MATHJAX_RELPATH "${mathjax_relpath}")
 # This functions adds the necessary routines for the generation of the
 # Doxyfile[.in] files needed to doxygen.
 macro(prepare_doxyfile)
-  cmake_parse_arguments(DOXYFILE "" "TARGET" "" ${ARGN})
+  cmake_parse_arguments(DOXYFILE "" "TARGET;DOXYFILE" "TAGFILES" ${ARGN})
 
   # default target name is the module name
   if(NOT DOXYFILE_TARGET)
@@ -125,22 +125,31 @@ macro(prepare_doxyfile)
 
   # check whether module has a Doxylocal file
   find_file(_DOXYLOCAL Doxylocal PATHS ${CMAKE_CURRENT_SOURCE_DIR} NO_DEFAULT_PATH)
-  # convert use_mathjax to YES/NO for Doxygen
-  set(use_mathjax $<IF:$<BOOL:${use_mathjax}>,YES,NO>)
+  set(make_doxyfile_options
+    -D DOT_TRUE=${DOT_TRUE}
+    -D DUNE_MOD_NAME=${PROJECT_NAME}
+    -D DUNE_MOD_VERSION=${ProjectVersion}
+    -D DOXYSTYLE=${DOXYSTYLE_FILE}
+    -D DOXYGENMACROS=${DOXYGENMACROS_FILE}
+    -D abs_top_srcdir=${CMAKE_SOURCE_DIR}
+    -D top_srcdir=${${PROJECT_NAME}_SOURCE_DIR}
+    -D DOXYGEN_TAGFILES=${DOXYFILE_TAGFILES}
+    -D DUNE_USE_MATHJAX=$<IF:$<BOOL:${use_mathjax}>,YES,NO>
+    -D DUNE_MATHJAX_RELPATH="${mathjax_relpath}"
+    -D DOXYFILE=${DOXYFILE_DOXYFILE}
+    -P ${scriptdir}/CreateDoxyFile.cmake)
   if(_DOXYLOCAL)
-    set(make_doxyfile_command ${CMAKE_COMMAND} -D DOT_TRUE=${DOT_TRUE} -D DUNE_MOD_NAME=${ProjectName} -D DUNE_MOD_VERSION=${ProjectVersion} -D DOXYSTYLE=${DOXYSTYLE_FILE} -D DOXYGENMACROS=${DOXYGENMACROS_FILE}  -D DOXYLOCAL=${CMAKE_CURRENT_SOURCE_DIR}/Doxylocal -D abs_top_srcdir=${CMAKE_SOURCE_DIR} -D srcdir=${CMAKE_CURRENT_SOURCE_DIR} -D top_srcdir=${${PROJECT_NAME}_SOURCE_DIR} -D DUNE_USE_MATHJAX="${use_mathjax}" -D DUNE_MATHJAX_RELPATH="${mathjax_relpath}" -P ${scriptdir}/CreateDoxyFile.cmake)
-    add_custom_command(OUTPUT Doxyfile.in Doxyfile
-      COMMAND ${make_doxyfile_command}
-      COMMENT "Creating Doxyfile.in"
+    add_custom_command(OUTPUT ${DOXYFILE_DOXYFILE}.in ${DOXYFILE_DOXYFILE}
+      COMMAND ${CMAKE_COMMAND} ${make_doxyfile_options} -D DOXYLOCAL=${CMAKE_CURRENT_SOURCE_DIR}/Doxylocal -D srcdir=${CMAKE_CURRENT_SOURCE_DIR} -P ${scriptdir}/CreateDoxyFile.cmake
+      COMMENT "Creating ${DOXYFILE_DOXYFILE}.in"
       DEPENDS ${DOXYSTYLE_FILE} ${DOXYGENMACROS_FILE} ${CMAKE_CURRENT_SOURCE_DIR}/Doxylocal)
   else()
-    set(make_doxyfile_command ${CMAKE_COMMAND} -D DOT_TRUE=${DOT_TRUE} -D DUNE_MOD_NAME=${ProjectName} -D DUNE_MOD_VERSION=${DUNE_MOD_VERSION} -D DOXYSTYLE=${DOXYSTYLE_FILE} -D DOXYGENMACROS=${DOXYGENMACROS_FILE} -D abs_top_srcdir=${CMAKE_SOURCE_DIR} -D top_srcdir=${${PROJECT_NAME}_SOURCE_DIR} -D DUNE_USE_MATHJAX="${use_mathjax}" -D DUNE_MATHJAX_RELPATH="${mathjax_relpath}" -P ${scriptdir}/CreateDoxyFile.cmake)
-    add_custom_command(OUTPUT Doxyfile.in Doxyfile
-      COMMAND ${make_doxyfile_command}
-      COMMENT "Creating Doxyfile.in"
+    add_custom_command(OUTPUT ${DOXYFILE_DOXYFILE}.in ${DOXYFILE_DOXYFILE}
+      COMMAND ${CMAKE_COMMAND}  ${make_doxyfile_options} -P ${scriptdir}/CreateDoxyFile.cmake
+      COMMENT "Creating ${DOXYFILE_DOXYFILE}.in"
       DEPENDS ${DOXYSTYLE_FILE} ${DOXYGENMACROS_FILE})
   endif()
-  add_custom_target(doxyfile_${DOXYFILE_TARGET} DEPENDS Doxyfile.in Doxyfile)
+  add_custom_target(doxyfile_${DOXYFILE_TARGET} DEPENDS ${DOXYFILE_DOXYFILE}.in ${DOXYFILE_DOXYFILE})
 endmacro(prepare_doxyfile)
 
 macro(add_doxygen_target)
@@ -154,11 +163,6 @@ macro(add_doxygen_target)
     set(DOXYGEN_TARGET ${PROJECT_NAME})
   endif()
 
-  # default output is html
-  if(NOT DOXYGEN_OUTPUT)
-    set(DOXYGEN_OUTPUT "${CMAKE_CURRENT_BINARY_DIR}/html")
-  endif()
-
   dune_module_path(MODULE dune-common RESULT scriptdir SCRIPT_DIR)
   if(PROJECT_NAME STREQUAL "dune-common")
     set(DOXYSTYLE_FILE ${CMAKE_CURRENT_SOURCE_DIR}/Doxystyle)
@@ -167,40 +171,50 @@ macro(add_doxygen_target)
   message(STATUS "Using scripts from ${scriptdir} for creating doxygen stuff.")
 
   if(TARGET Doxygen::doxygen)
-    prepare_doxyfile(TARGET ${DOXYGEN_TARGET})
+
+    set(DOXYGEN_BUILD_TAGFILES ${DOXYGEN_TAGFILES})
+    foreach(module ${DUNE_FOUND_DEPENDENCIES})
+      set(DOXYGEN_BUILD_TAGFILES "${DOXYGEN_BUILD_TAGFILES} ${${module}_DOXYGEN_DIR}/${module}.tag=${${module}_DOXYGEN_DIR}/html")
+    endforeach()
+    prepare_doxyfile(TARGET ${DOXYGEN_TARGET}_build TAGFILES ${DOXYGEN_BUILD_TAGFILES} DOXYFILE Doxyfile)
     # custom command that executes doxygen
-    add_custom_command(OUTPUT ${DOXYGEN_OUTPUT}
-      COMMAND ${CMAKE_COMMAND} -D DOXYGEN_EXECUTABLE=$<TARGET_FILE:Doxygen::doxygen> -P ${scriptdir}/RunDoxygen.cmake
+    add_custom_command(OUTPUT ${DOXYGEN_OUTPUT} html/ ${PROJECT_NAME}.tag
+      COMMAND ${CMAKE_COMMAND} -D DOXYGEN_EXECUTABLE=$<TARGET_FILE:Doxygen::doxygen> -D DOXYFILE=Doxyfile -P ${scriptdir}/RunDoxygen.cmake
       COMMENT "Building doxygen documentation. This may take a while"
-      DEPENDS Doxyfile.in ${DOXYGEN_DEPENDS})
+      DEPENDS Doxyfile ${DOXYGEN_DEPENDS})
     # Create a target for building the doxygen documentation of a module,
     # that is run during make doc
-    add_custom_target(doxygen_${DOXYGEN_TARGET}
-      DEPENDS ${DOXYGEN_OUTPUT})
-    add_dependencies(doxygen doxygen_${DOXYGEN_TARGET})
+    add_custom_target(doxygen_${DOXYGEN_TARGET}_build
+      DEPENDS html ${PROJECT_NAME}.tag)
+    add_dependencies(doxygen doxygen_${DOXYGEN_TARGET}_build)
+
+    set_property(GLOBAL PROPERTY ${PROJECT_NAME}_DOXYGEN_DIR "${CMAKE_CURRENT_BINARY_DIR}")
 
     # Use a cmake call to install the doxygen documentation and create a
     # target for it
     include(GNUInstallDirs)
+
+    set(DOXYGEN_INSTALL_TAGFILES ${DOXYGEN_TAGFILES})
+    foreach(module ${DUNE_FOUND_DEPENDENCIES})
+      # TODO: Fix proper doxygen directory structure. This assumes that the tagfiles are installed in the same relative location as the build tagfiles.
+      set(DOXYGEN_INSTALL_TAGFILES "${DOXYGEN_INSTALL_TAGFILES} ${${module}_DOXYGEN_DIR}/installdir/${module}.tag=../../../${module}/doxygen/html")
+    endforeach()
+
+    prepare_doxyfile(TARGET ${DOXYGEN_TARGET}_install TAGFILES ${DOXYGEN_INSTALL_TAGFILES} DOXYFILE installdir/Doxyfile)
+    # custom command that executes doxygen
+    add_custom_command(OUTPUT ${DOXYGEN_OUTPUT} installdir/html/ installdir/${PROJECT_NAME}.tag
+      COMMAND ${CMAKE_COMMAND} -D DOXYGEN_EXECUTABLE=$<TARGET_FILE:Doxygen::doxygen> -D DOXYFILE=Doxyfile -P ${scriptdir}/RunDoxygen.cmake
+      COMMENT "Building doxygen documentation. This may take a while"
+      WORKING_DIRECTORY installdir
+      DEPENDS installdir/Doxyfile ${DOXYGEN_DEPENDS})
+
+    add_custom_target(doxygen_${DOXYGEN_TARGET}_install
+      DEPENDS installdir/html/ installdir/${PROJECT_NAME}.tag)
+
     # When installing call cmake install with the above install target
     install(CODE
-      "execute_process(COMMAND ${CMAKE_COMMAND} --build ${CMAKE_BINARY_DIR} --target doxygen_${ProjectName}
-          WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR})
-        file(GLOB doxygenfiles
-          GLOB ${CMAKE_CURRENT_BINARY_DIR}/html/*.html
-          ${CMAKE_CURRENT_BINARY_DIR}/html/*.js
-          ${CMAKE_CURRENT_BINARY_DIR}/html/*.png
-          ${CMAKE_CURRENT_BINARY_DIR}/html/*.css
-          ${CMAKE_CURRENT_BINARY_DIR}/html/*.gif
-          ${CMAKE_CURRENT_BINARY_DIR}/*.tag
-          )
-        set(doxygenfiles \"\${doxygenfiles}\")
-        foreach(_file \${doxygenfiles})
-           get_filename_component(_basename \${_file} NAME)
-           # Manifest is generated when prefix was set at configuration time, otherwise is skipped
-           LIST(APPEND CMAKE_INSTALL_MANIFEST_FILES ${CMAKE_INSTALL_FULL_DOCDIR}/doxygen/\${_basename})
-         endforeach()
-         file(INSTALL \${doxygenfiles} DESTINATION ${CMAKE_INSTALL_DOCDIR}/doxygen)
-         message(STATUS \"Installed doxygen into ${CMAKE_INSTALL_DOCDIR}/doxygen\")")
+      "execute_process(COMMAND ${CMAKE_COMMAND} --build ${CMAKE_BINARY_DIR} --target doxygen_${ProjectName}_install WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR})")
+    install(DIRECTORY "${CMAKE_CURRENT_BINARY_DIR}/installdir/html/" DESTINATION ${CMAKE_INSTALL_DOCDIR}/doxygen/html)
+    install(FILES "${CMAKE_CURRENT_BINARY_DIR}/installdir/${PROJECT_NAME}.tag" DESTINATION ${CMAKE_INSTALL_DOCDIR}/doxygen)
   endif()
 endmacro(add_doxygen_target)